LargeUpdate.js
Wrote a TON of code today.  Started this morning by trying to do this project using just the teq ive learned so far.  I took that and translated it into OOP.  Still trying to grasp the understanding 100%, but im getting there!

created : player class and game class.

I used inheritence with my computer class which is a child of parent player

I finished up a switchCase which lets the user choose their gesture.

I finished an array that includes all the gestures, which the random computer function pulls one and returns a random.

finished displayed rules

finished user name prompt (not necessary, but it will add a custom touch throughtout with the users name.