If/Else.js
finished all the statements within the function.

Its not completly done, I need to test and reach out to make sure im going in the correct direction.

my next step is to create a forLoop around the function that does the math to check for a winner that way it will be able to loop around it to play best out of three. it will also keep track of each players score.