letsDoTheMathToFindAWinner.js
function has been completly overhauled.  I had to change around everything as my code was written wrong and things were not making sense.  I tested and it seems all is working with that. so as of right now I have best out of 1 working for single and multi.

the next step will be to add a while loop once more (in the correct location this time) so that I can keep score and make the game best out of three.